refactor(api): consolidate drizzle schemas into one file
