+++ apps/api/src/server/services/AuthService.ts
@@ -6,7 +6,7 @@ import { pbkdf2, randomBytes, timingSafeEqual } from 'node:crypto';
 
 import { isSqlError } from '../utils';
 import { BaseService, type Resources } from './Service';
-import { user } from '../../models/user';
+import { user } from '../../models/schema';
 
 const pbkdf2Promise = promisify(pbkdf2);
